Rename get_item and save_item for consistency
The API implements the CRUD operations so it makes more sense
to have read_item instead of get_item, to go with create_item
and update_item.

Also added an explicit modify_item method to the storage module
instead of using save_item to both save a newly created item and
to change an existing item. At some point there will be a need to
distinguish between the two without doing an additional lookup.